ターゲットを追加

+++ OpenGL_Game/src/bullet/control_bullet.cpp
@@ -26,7 +26,7 @@ void Update(Vector3 camera_place, Vector3 camera_viewpoint,
 		double camera_speed) {
 	//登録
 	if (input::get_mouse_left_button_frame() == 1) {
-		if (bullet_.size() < kMaxBulletNum) { //最大数を下回っている
+		if ((int)bullet_.size() < kMaxBulletNum) { //最大数を下回っている
 			//登録場所を少し上にずらす
 			camera_viewpoint.y += 0.5;
 			//毎フレームの移動角度(3座標)を計算
